Update to tutorial 1.2 - examples only to LLO
In this update I have changed the mention of hdata to ldata to make it more obvious we are dealing with LLO data. In the second example which looks at a longer stretch of data I have also changed this to look at LLO data rather than LHO data. On in the challenge section with the reader now download that data.